makefile ?

+++ ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: ajuliao- <ajuliao-@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/14 12:08:00 by ajuliao-          #+#    #+#             */
-/*   Updated: 2023/11/07 23:23:24 by ajuliao-         ###   ########.fr       */
+/*   Updated: 2023/11/09 19:21:00 by ajuliao-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
-	int	len;
+	size_t	len;
 
 	len = 0;
 	while (str[len] != '\0')
